% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTableWidget.R
\name{dataTableWidget}
\alias{dataTableWidget}
\title{dataTableWidget}
\value{
shiny code which, wnen invoked (almost always by the shinyApp function, returns html

nothing

nothing

nothing

nothing

nothing
display each cell in each row as a single line of text, or multiple

nothing
}
\description{
an R6 class providing clean access to a simple html message box.
}
\details{
R6 Class representing a simple text display widget
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tableSelection}}{a reactive value, see server() below}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{dataTableWidget$new()}}
\item \href{#method-ui}{\code{dataTableWidget$ui()}}
\item \href{#method-server}{\code{dataTableWidget$server()}}
\item \href{#method-setTable}{\code{dataTableWidget$setTable()}}
\item \href{#method-selectRows}{\code{dataTableWidget$selectRows()}}
\item \href{#method-clearSelection}{\code{dataTableWidget$clearSelection()}}
\item \href{#method-selectColumns}{\code{dataTableWidget$selectColumns()}}
\item \href{#method-setWrapLinesOption}{\code{dataTableWidget$setWrapLinesOption()}}
\item \href{#method-clone}{\code{dataTableWidget$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new dataTableWidget.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableWidget$new(
  id,
  tbl,
  width = "95\%",
  height = "500px",
  border = "0px;",
  pageLength = 10,
  wrapLines = FALSE,
  selectionMode = "single",
  selectionTarget = "row",
  lengthMenu = c(5, 10, 25, 50),
  columnWidths = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the html document div id}

\item{\code{tbl}}{data.frame}

\item{\code{width}}{character, "800px;" by default.  "100\%;" can be useful.}

\item{\code{height}}{integer, "600px;" by default, percentage heights do not work.}

\item{\code{border}}{string, "0px;" by default, "1px blue solid; border-radius: 5px;" useful.}

\item{\code{pageLength}}{integer, 10 by default}

\item{\code{wrapLines}}{logical, FALSE by default, gives more rows on screen}

\item{\code{selectionMode}}{character, "single", "none", or "multiple", "single" by default}

\item{\code{selectionTarget}}{character, "row" or "column" or "cell"}

\item{\code{lengthMenu}}{integers specifies page length options (number of rows onscreen). default (5,10,25,50)}

\item{\code{columnWidths}}{integer vector, widths in pixels, NULL default, applies all
supplied, default for the remainder}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `dataTableWidget` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ui"></a>}}
\if{latex}{\out{\hypertarget{method-ui}{}}}
\subsection{Method \code{ui()}}{
defines the html structure of this widget
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableWidget$ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-server"></a>}}
\if{latex}{\out{\hypertarget{method-server}{}}}
\subsection{Method \code{server()}}{
defines the event handlers used with the ui
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableWidget$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{list, managed by Shiny}

\item{\code{output}}{list, managed by Shiny}

\item{\code{session}}{list (environment?) managed by Shiny}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setTable"></a>}}
\if{latex}{\out{\hypertarget{method-setTable}{}}}
\subsection{Method \code{setTable()}}{
installs a new data.frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableWidget$setTable(tbl)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tbl}}{a data.frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-selectRows"></a>}}
\if{latex}{\out{\hypertarget{method-selectRows}{}}}
\subsection{Method \code{selectRows()}}{
highlight the rows, described by their ordingal number.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableWidget$selectRows(rowNumbers)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rowNumbers}}{an integer vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clearSelection"></a>}}
\if{latex}{\out{\hypertarget{method-clearSelection}{}}}
\subsection{Method \code{clearSelection()}}{
clears the current selection, if any
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableWidget$clearSelection()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-selectColumns"></a>}}
\if{latex}{\out{\hypertarget{method-selectColumns}{}}}
\subsection{Method \code{selectColumns()}}{
select columns in the table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableWidget$selectColumns(colNumbers)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{colNumbers}}{numeric vector, which columns to select}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setWrapLinesOption"></a>}}
\if{latex}{\out{\hypertarget{method-setWrapLinesOption}{}}}
\subsection{Method \code{setWrapLinesOption()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableWidget$setWrapLinesOption(wrapLines)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wrapLines}}{logical, TRUE or FALSE}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataTableWidget$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
